+++ HISTORY.md
@@ -45,6 +45,7 @@
 ### 2022-10-21
 * objects: enabled same-named global and sprite-local variables
 * gui: incremented dev version to 8.1
+* threads: tweaked context binding
 
 ### 2022-10-20
 * removed PIPE from the iteration-composition library